7-DashboardSIG/backend/requirements.txt
# DashboardSIG Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
kafka-python==2.0.2
pydantic==2.5.3
reportlab==4.0.8
pandas==2.1.4
python-multipart==0.0.6
httpx==0.26.0
